requirements.txt
streamlit
requests
beautifulsoup4
pandas
finnhub-python
gspread
oauth2client